Fixed boolean fields

+++ module/VuFind/sql/mysql.sql
@@ -138,7 +138,7 @@ CREATE TABLE `search` (
   `folder_id` int(11) DEFAULT NULL,
   `created` datetime NOT NULL DEFAULT '2000-01-01 00:00:00',
   `title` varchar(20) DEFAULT NULL,
-  `saved` int(1) NOT NULL DEFAULT '0',
+  `saved` tinyint(1) NOT NULL DEFAULT '0',
   `search_object` blob,
   `checksum` int(11) DEFAULT NULL,
   `notification_frequency` int(11) NOT NULL DEFAULT '0',
@@ -265,7 +265,7 @@ CREATE TABLE `user_list` (
   `title` varchar(200) NOT NULL,
   `description` text,
   `created` datetime NOT NULL DEFAULT '2000-01-01 00:00:00',
-  `public` tinyint NOT NULL DEFAULT '0',
+  `public` tinyint(1) NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `user_id` (`user_id`),
   CONSTRAINT `user_list_ibfk_1` FOREIGN KEY (`user_id`) REFERENCES `user` (`id`) ON DELETE CASCADE
